ft(Folder): folder restructuring
[Finishes: #174486675]
