+++ src/route/App.js
@@ -1,17 +1,17 @@
 import React from "react"
 import { Route, Switch, BrowserRouter } from "react-router-dom"
-import Login from "../views/Login"
-import LandingPage from "../views/LandingPage"
-import Signup from "../views/SignUp"
+import Login from "../views/login/Login.js"
+import LandingPage from "../views/landingPage/LandingPage.js"
+import Signup from "../views/signUp/SignUp.js"
 import ResetPassword from "../views/ResetPassword"
-import UpdateProfile from "../views/UpdateProfile"
-import Home from "../views/Home"
+import UpdateProfile from "../views/profile/UpdateProfile"
+import Home from "../views/viewProduct/Home.js"
 import ViewOneProduct from "../views/ViewOneProduct"
 import LoanAgreement from "../views/LoanAgreementPage"
-import Loans from "../views/Loans"
+import Loans from "../views/viewMyLoans/Loans"
 import Pay from "../views/Pay"
 import Seller from "../views/Seller"
-import AddProduct from "../views/AddProduct"
+import AddProduct from "../views/addProduct/AddProduct"
 import ViewRequest from "../views/ViewRequest"
 import OpenRequest from "../views/OpenRequest"
 import classes from "./App.css"
